requirements.txt
Flask
Flask-JWT-Extended
Flask-Cors
Flask-Bcrypt
Flask-SQLAlchemy
cloudinary
python-dotenv
Werkzeug
pytest
